+++ app/src/main/java/cat/tecnocampus/mobileapps/practica2/HomarMasachsFrancescMeninoSuredaPau/Entities/Player.java
@@ -18,7 +18,7 @@ public class Player {
     public int games;
 
     @ColumnInfo(name = "total_points")
-    public int totalPoints;
+    public double totalPoints;
 
     public void setNickname(String nickname) {
         this.nickname = nickname;
@@ -28,7 +28,7 @@ public class Player {
         this.games = games;
     }
 
-    public void setTotalPoints(int totalPoints) {
+    public void setTotalPoints(double totalPoints) {
         this.totalPoints = totalPoints;
     }
 
@@ -41,7 +41,7 @@ public class Player {
         return games;
     }
 
-    public int getTotalPoints() {
+    public double getTotalPoints() {
         return totalPoints;
     }
 }
